Update focusResource if needed when wl_keyboard is bound
If a QtWayland compositor gives focus to a surface right after its
creation during the client startup, the keyboard resource may not be
bound yet by the client.
In this case, the surface is correctly marked as focused, but the
keyboard resource is never marked as focused, and thus no keys are
ever sent to the client.

To fix this, the focusResource is updated if needed after wl_keyboard
is bound.

This can be reproduced with weston-simple-im (patched to use wl_shell
instead of xdg_shell) and qml-compositor, modified to enable
TextInputExtension and call takeFocus() at the end of windowAdded().

Change-Id: I551cb5bc56c05a1e5187b23108f4ef80468782dc
Reviewed-by: Giulio Camuffo <giulio.camuffo@jollamobile.com>