Workaround for qwindow-compositor on platforms other than xcb
Creating the QWaylandOutput after the compositor's create() posts
an event to the output object and defers initialization. This is
fragile since there is nothing guaranteeing the application will not
get to render in the meantime and not start calling functions like
defaultOutput(), which all return null since there are 0 outputs
registered at that time due to the deferred QWaylandOutput::initialize().

Until the issue gets solved by redesigning the behavior somehow, swap
the two lines to get qwindow-compositor running with eglfs. This way
the output's initialize() is called directly from the compositor's
create() and this all output-related functions are usable afterwards
immediately.

Change-Id: Ie4203d0a4d0ec35804ef8c320121750f02b50724
Reviewed-by: Paul Olav Tvete <paul.tvete@theqtcompany.com>