Pure QML example: add shortcut to quit
On the eglfs platform, there is no obvious way to exit back to a console, so
it's too easy to get stuck in this example and need to reboot via ssh or some
such.  Ctrl-Alt-Backspace is the historical way of exiting from an X11 or
Weston session, so we use it here.

Change-Id: Ibe93acb0a81ef47384d5449760f25ad6e4cac8d8
Reviewed-by: Pier Luigi Fiorini <pierluigi.fiorini@gmail.com>