Add popup support to QWaylandQuickCompositor with the wl shell.
When the shell surface emits the QWaylandShell::setPopup signal the
QWaylandQuickWlShellSurfaceItem will move it self into its popup-parent,
and also remove all transforms (rotation, scaling, etc.), our new
parent already have the same transforms set.

During the time a popup is active it installs a filter
which handles the closing of popups if the user clicks
outside the popup. This requires that the background item of the
compositor accepts mouse press.

Change-Id: I3e4c3c8728795d4b38a30a9bce178dc315643de2
Reviewed-by: Johan Helsing <johan.helsing@qt.io>
Reviewed-by: Paul Olav Tvete <paul.tvete@theqtcompany.com>
Reviewed-by: Giulio Camuffo <giulio.camuffo@kdab.com>