qwindow-compositor: Fix cursor not being set for mapped surfaces
If a client sent a set_cursor request with a mapped surface,
qwindow-compositor did not call updateCursor until the next time the
surface was committed, which resulted in cursors with incorrect pixmaps
and hotspots.

This is solved by adding a check to see if the surface in set_cursor is
already mapped, and if so, update the cursor immediately.

The issue did not affect Qt clients, since commits are sent after
set_cursor.

Change-Id: I8b2555b68e9517e04f5bb8c5affe09d84bc97c0c
Reviewed-by: Pier Luigi Fiorini <pierluigi.fiorini@gmail.com>
Reviewed-by: Erik Larsson <erik@ortogonal.com>
Reviewed-by: Paul Olav Tvete <paul.tvete@theqtcompany.com>