Create a fallback shellSurface in wl_shell::get_shell_surface
This makes QWaylandShellPrivate::shell_get_shell_surface behave
similarly to QWaylandCompositorPrivate::compositor_create_surface.

First, a signal is emitted letting a compositor implementation create
its own subclass of QWaylandShellSurface if desired. If it's not
created, a plain QWaylandShellSurface is created instead.

It also makes the API more consistent, as the creation of shell surfaces
now mimick the creation of surfaces and xdg surfaces.

qwindow-compositor has been updated to take advantage of the new API

Change-Id: I7c7262dd253dc37843e8e471a4ead5f6e87b0ced
Reviewed-by: Giulio Camuffo <giulio.camuffo@kdab.com>