Set state to expected state inside compositor
Some input methods like qtvirtualkeyboard expect to receive all updates
to surrounding text, cursor position, etc before processing of the
QInputMethodEvent is done. Else they assume the updates are happended because
of some external input and reset all internal state regarding the input.

Since we send the text events first to the client before we get status updates
back from the client, the status of the virtual keyboard gets
constantly reset.

To work around this issue, we already update the state in the compositor
while processing QInputMethodEvents to the state we expect to get sent
later from the clients. As long as there is no difference between the
state we expect the client will have and the precalculated state there
will be no status updates sent to the qtvirtualkeyboard outside of processing
QInputMethodEvents.

Change-Id: Ife1b541510d754657fc18281a8d765ea97935fcc
Reviewed-by: Giulio Camuffo <giulio.camuffo@kdab.com>