Emit a signal when QWaylandXdgSurface changes window type
This allows compositors to track when a window changes type and
act accordingly (for example do a different show animation).

Compositors might also want to keep track of the window type in
order to filter out windows when presenting them to the user,
think of a spread windows effect that only shows top level windows.

Change-Id: I13c0ed3db2bdc6318ff1f4345d065e188b7c2435
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>