Fix drag-and-drop for QML compositors
Dragging within the same client was broken because handleStartDrag called
setInputEventsEnabled(false), which meant the surface was ignored as a target
for the drag.

With this change, input events are not disabled, and wl_data_device.motion
events are sent to the source surface like they should. Also wl_pointer.leave
and enter are sent before and after the drag.

Change-Id: Ife46ca3a9d44a660a5cf029093842adb0e3eca7c
Reviewed-by: Giulio Camuffo <giulio.camuffo@kdab.com>