multi-screen example: add control-alt-backspace Shortcut to quit
Same as 8cdd72afc11d96ec41f65577b38ffea63bc7294c on the pure-qml example.
On the eglfs platform, there is no obvious way to exit back to a console, so
it's too easy to get stuck in this example and need to reboot via ssh or some
such.  Ctrl-Alt-Backspace is the historical way of exiting from an X11 or
Weston session, so we use it here.

Change-Id: Ife3318675f1b5dbe640feb8f3f99e80efbd5dcfd
Reviewed-by: Pier Luigi Fiorini <pierluigi.fiorini@liri.io>
Reviewed-by: Johan Helsing <johan.helsing@qt.io>