Set window screen from wl_surface.enter and leave events
Removes the pointer mScreen, which would previously cause a crash if the screen
was removed.

Ensures that QWindow::screen() is correct, except in the cases where:

- The compositor has not yet sent enter and leave events (in which case the
  primary output is returned).
- The compositor is not sending enter and leave events (although this is
  mandatory, some compositors don't do this).
- The application developer has tried to move the window with
  QWindow::setScreen(QScreen *). Since there is no way for a client to ask to
  be moved to a specific monitor in windowed mode, we return the requested
  screen until a new enter or leave event is received.

This will also be useful when implementing/fixing features where the current
screen matters. Examples are QT_AUTO_SCREEN_SCALE_FACTOR and the optional
output parameter to wl_shell_surface.set_fullscreen.

Task-number: QTBUG-62044
Change-Id: Iafde2e278fbc8876e8dafe5b2a4d2482fdc7961a
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>