Add client parameter for set_cursor
Identify the sender of set_cursor request to determine
surfaces that the cursor surface is associated with.

New signal cursorSurfaceRequested is added.
And the old signal cursorSurfaceRequest is deprecated.

A quote from the wayland protocol spec -
"The cursor actually changes only if the pointer focus for this device
is one of the requesting client's surfaces or the surface parameter is
the current pointer surface."

Pick-to: 6.1
Change-Id: I1cd1345d6479877925c200859868ae8717afd1fa
Reviewed-by: Elvis Lee <kwangwoong.lee@lge.com>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>