Compositor tests: Don't assume global numeric name == wayland object id
Throughout the code we consistently and confusingly use the term, "id", instead
of "numeric name" for Wayland globals. This is confusing because it's not the
same as thing as wl_proxy_get_id returns, although the values frequently are
the same because we typically bind to all globals in the order they are
announced.

So in this patch we turn m_outputs into a map with the numeric name and compare
it with the numeric name we get in handleGlobalRemove.

Change-Id: I0716bf66e12dbbf7b05b9ca5faa5d2beba297af5
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>