ivi: Try the ivi-shell as shell integration as well
When using the IviApplication in a compositor and launching a
Qt Client it will fail to initialize a shell. This is because
QtWaylandClient::QWaylandIntegration::initializeShellIntegration
will only try some shells by default. Add the ivi-application
to make it work out of the box. Add it last to have XDG and WL
shell take preference and avoid loading the libivi-shell.so.

Change-Id: I5c97c65d81434cba59cf9cb5bbe4b6fd8ccf4757
Reviewed-by: Johan Helsing <johan.helsing@qt.io>
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>