Compositor API: Don't leak OpenGL textures when running multi-threaded
If a client destroyed buffers, they would not get deleted because the event
would be handled on the GUI thread without a valid OpenGL context. I.e: you
would see the following output:

    QOpenGLTexturePrivate::destroy() called without a current context.
    Texture has not been destroyed

Work around it by keeping a list of buffers that have been orphaned and delete
them whenever we have a valid OpenGL context.

Change-Id: I31c7c4cff179bd68cd954714b2a8e3e2a74675b2
Reviewed-by: Dominik Holland <dominik.holland@pelagicore.com>
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>