Add readme for internal protocol extension folder
There have been some misunderstandings about the compatibility guarantees
of our internal protocols. I.e. some compositors have started
implementing them, and then we have removed or made backwards
incompatible changes to the protocols.

Also, when we've made backwards incompatible changes, we've previously
incorrectly used the version attribute of the interface, however that
attribute is supposed to be used for backwards *compatible* changes. This
causes problems for both ourselves and others.

So let's be explicit about it in a README and be more careful when
reviewing protocol extension changes if they don't bump the version
suffix (i.e. rename the protocol).

This is the same versioning and naming scheme as used by the unstable
protocols in wayland-protocols.

Task-number QTBUG-68423

Change-Id: I9c1a7e13c9356660914bbfeff4139ec033f1d443
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>