Client (xdg-shell): Only restore normal size when not maximized
kwin and perhaps other compositors sometimes send questionable configure events
with state set to maximized and size == {0,0}.

Previously, we would then restore the m_normalSize, which would give us a
maximized window with the size of a windowed (normal) one.

A size of zero usually means that it's up to the client to decide, which makes
sense for non-maximized windows, but not so much for maximized ones. This is
what the protocol spec says about the maximized state:

    The surface is maximized. The window geometry specified in the configure
    event must be obeyed by the client.

It's fixed in the dev version of kwin, but until then, let's be on the safe
side and only resize the window if things make sense.

[ChangeLog][QPA plugin] Fixed a bug where maximized windows would resize to
their unmaximized size if the compositor sent an invalid configure event.

Change-Id: I2371b29c82426ac48cd1c18c14c3dd0fe4f2250e
Reviewed-by: David Edmundson <davidedmundson@kde.org>
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>