Compositor: Update WaylandQuickItem size on new dpr or scaleFactor
updateSize() depends on both screen devicePixelRatio and output scaleFactor.
Make sure to call it when they change.

This still doesn't handle all cases, because there is no signal for when a
screen device pixel ratio changes, but we update when we get a new screen,
which should be good enough in most cases

Change-Id: I4550cf580d4821e4b0f4da0d2eb7bb8d9ab823cf
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>