Client: Decrease buffer_scale for small cursor themes
Not all setups or themes have cursors with a matching DPI. In such cases,
wl_cursor_load_theme will then return a theme that is the closest resolution it
can get.

With the previous implementation, cursors themes without a high dpi version
would become become really tiny on high DPI displays.

This patch prevents it by setting a lower wl_surface.scale for those themes.

This also implements proper tracking of the cursor surface's entered outputs
(i.e. if the entered surface is destroyed, the scale is reset, and similarly
the following sequence of events should also be handled:

    wl_surface.enter(wl_output@1)
    wl_surface.enter(wl_output@2)
    wl_surface.leave(wl_output@2)

In the old implementation, we would be stuck with the scale from wl_output@2,
but now we now should correctly get the scale of wl_output@1.

[ChangeLog][QPA plugin] Cursors on high DPI screens are now scaled up if the
cursor theme does not have an appropriate high resolution version.

Change-Id: Ic87d00e35612b5afdf8c2e3a4463fcfef1f1f09d
Reviewed-by: Giulio Camuffo <giulio.camuffo@kdab.com>