Compositor: Respect wl_shm format when converting to a QImage
[ChangeLog][Compositor] Fixed a bug where compositors would advertise
support for numerous wl_shm pixel formats and then blindly assume
everything to be ARGB32_Premultiplied.

Note that this may cause expensive conversions if the format of the client
and the screen differs. The solution is probably to add API for letting the
compositor opt in to the formats it wants to support (except for argb8888
and xrgb8888 which are mandatory).

Fixes: QTBUG-75635
Change-Id: I07ca3dd4ef9633222d53361860fc0ab99f02eae7
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>