QWaylandXdgShellV5: optimize closeAllPopups()
The code iterated over QMultiMap::keys(), calling QMultiMap::value()
for each key, reversed the list returned by values() and then iterated
over it to call sendPopupDone().

The QMultiMap is keyed on a pointer type, so the order of keys is
unspecified. We can therefore simply iterate backwards over the
QMultiMap, and just call our function that way. The order in which the
keys are visited is reversed, but since they are unordered, this does
not change behavior. Since we iterate backwards, we will visit popups
for the same key in the same order the old code did.

Except we don't create 1 + N new containers in doing so.

Only problem: well-maintained QMultiMap doesn't have rbegin()/rend(),
yet, so roll our own std::make_reverse_iterator for easier porting
later.

Change-Id: Ibec35cb4262dd5bb19c74821e85baa956f67dd05
Reviewed-by: Johan Helsing <johan.helsing@qt.io>